requirements.txt
python-dotenv
openai
langchain
langchain-openai
atlassian-python-api
uuid
tiktoken
lxml
request
requests_oauthlib
langchain_experimental
langchain-text-splitters
pypdf
fast
jsonschema
logging
httpx
azure-identity
azure-cosmos
docx2txt
